requirements.txt
# Core ML dependencies
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Hyperparameter tuning
optuna>=3.0.0

# Imbalanced data handling
imbalanced-learn>=0.11.0

# Explanations
shap>=0.42.0
lime>=0.2.0

# Fairness
fairlearn>=0.9.0

# Visualizations
plotly>=5.0.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Model serialization and deployment
joblib>=1.3.0
onnx>=1.14.0
skl2onnx>=1.15.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Statistical tests
scipy>=1.10.0

# Optional dependencies (commented out by default)
# click>=8.0  # For CLI
# streamlit>=1.0.0  # For dashboard
# fastapi>=0.100.0  # For API serving
# uvicorn>=0.23.0  # For API serving